+++ crates/ruma-signatures/src/lib.rs
@@ -62,21 +62,22 @@
 pub use ruma_common::{IdParseError, SigningKeyAlgorithm};
 
 pub use self::{
+    ed25519::{Ed25519KeyPair, Ed25519KeyPairParseError, Ed25519VerificationError},
     error::{JsonError, VerificationError},
-    functions::{
-        canonical_json, content_hash, hash_and_sign_event, reference_hash, sign_json,
-        verify_canonical_json_bytes, verify_event, verify_json,
+    hash::{content_hash, reference_hash},
+    sign::{KeyPair, Signature, hash_and_sign_event, sign_json},
+    verify::{
+        PublicKeyMap, PublicKeySet, Verified, required_server_signatures_to_verify_event,
+        to_canonical_json_string_for_signing, verify_canonical_json_bytes, verify_event,
+        verify_json,
     },
-    keys::{Ed25519KeyPair, Ed25519KeyPairParseError, KeyPair, PublicKeyMap, PublicKeySet},
-    signatures::Signature,
-    verification::{Ed25519VerificationError, Verified},
 };
 
+mod ed25519;
 mod error;
-mod functions;
-mod keys;
-mod signatures;
-mod verification;
+mod hash;
+mod sign;
+mod verify;
 
 #[cfg(test)]
 mod tests {
@@ -90,7 +91,8 @@ mod tests {
     use serde_json::{from_str as from_json_str, to_string as to_json_string};
 
     use super::{
-        Ed25519KeyPair, canonical_json, hash_and_sign_event, sign_json, verify_event, verify_json,
+        Ed25519KeyPair, hash_and_sign_event, sign_json, to_canonical_json_string_for_signing,
+        verify_event, verify_json,
     };
 
     fn pkcs8() -> Vec<u8> {
@@ -110,7 +112,7 @@ mod tests {
     /// Convenience for converting a string of JSON into its canonical form.
     fn test_canonical_json(input: &str) -> String {
         let object = from_json_str(input).unwrap();
-        canonical_json(&object).unwrap()
+        to_canonical_json_string_for_signing(&object).unwrap()
     }
 
     #[test]