+++ crates/ruma-signatures/CHANGELOG.md
@@ -25,7 +25,7 @@ Breaking changes:
     `Ed25519VerificationError::SignatureVerification`.
 - `Error::PduSize` is now `JsonError::PduTooLarge` allowing the following
   functions to return `JsonError` as an error type:
-  - `canonical_json()`
+  - `to_canonical_json_string_for_signing()`
   - `reference_hash()`
   - `content_hash()`
   - `sign_json()`
@@ -38,11 +38,20 @@ Breaking changes:
 - When verifying the signatures on a JSON object, signatures of keys that are
   not in the key map are ignored rather than returning an error. The
   `VerificationError::PublicKeyNotFound` variant was removed.
+- `Ed25519KeyPair` was moved under the new `ed25519` module with its error
+  types.
+- `Signature::new()` takes an `OwnedSigningKeyId<AnyKeyName>` and a `Vec<u8>`
+  and is now infallible.
+- `canonical_json()` was renamed to `to_canonical_json_string_for_signing()` to
+  clarify that is is not to be used outside of the signing/verifying context.
 
 Improvements:
 
 - Get a better error message when verifying a signature with a public key that
   has the wrong length.
+- Make `required_server_signatures_to_verify_event()` public, for homeservers to
+  get the list of servers whose public keys they need to provide to
+  `verify_event()`.
 
 # 0.19.0